chore(frontend): consolidate formatBytes to src/utils/formatters
