+++ src/components/saves/SaveFileRow.tsx
@@ -8,8 +8,8 @@ import { createElement } from "react";
 import { DialogButton } from "@decky/ui";
 import type { SaveFileStatus, SyncConflict } from "../../types";
 import { scrollFocusedToCenter } from "../../utils/scrollHelpers";
-import { formatTimestamp } from "../../utils/formatters";
-import { formatAttributionSegment, formatBytes, formatRelativeTime, pickLastSyncer, statusLabel } from "./helpers";
+import { formatBytes, formatTimestamp } from "../../utils/formatters";
+import { formatAttributionSegment, formatRelativeTime, pickLastSyncer, statusLabel } from "./helpers";
 
 // Label column width — keeps values aligned vertically across rows
 const LABEL_WIDTH = "88px";
